+++ src/ChoyUtilities/Assets/ChoyUtilities/Common/Runtime/Audio/ReserveAudioManager.cs
@@ -12,6 +12,7 @@
 //    See the License for the specific language governing permissions and
 //    limitations under the License.
 
+using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Unity.Mathematics;
@@ -21,7 +22,7 @@ using UnityEngine.Pool;
 
 namespace ChoyUtilities {
 
-    public class ReserveAudioManager : GenericSingleton<ReserveAudioManager>, IBootloader {
+    public partial class ReserveAudioManager : GenericSingleton<ReserveAudioManager>, IBootloader {
 
         public enum EAudioPriority : byte {
             Highest = 0,
@@ -36,7 +37,7 @@ namespace ChoyUtilities {
             Lowest = byte.MaxValue
         }
         
-        [SerializeField] protected Attribute attribute;
+        [SerializeField] protected ReserveAudioAttribute reserveAudioAttribute;
         [SerializeField] protected AudioSource audioSourcePrefab;
         [SerializeField] protected byte poolCount = 32;
         [SerializeField] protected bool loop;
@@ -50,7 +51,12 @@ namespace ChoyUtilities {
         protected List<int> PauseIndexes;
 
         protected ObjectPool<AudioSource> Pool;
-        
+
+        private void Start() {
+            if (BootLoader.Instance is null)
+                _ = Init();
+        }
+
         public async Task Init() {
             await Awaitable.NextFrameAsync(Token);
 
@@ -80,83 +86,7 @@ namespace ChoyUtilities {
             ListPool<int>.Release(PauseIndexes);
             await Awaitable.EndOfFrameAsync(Token);
         }
-        
-        public (float, AudioSource) PlayClipAtPos(string id, float3 pos, byte audioPriority = (byte)EAudioPriority.Average) {
-            return !attribute.pools.TryGet(id, out var resource) 
-                ? (0f, null) : PlayClipAtPos(resource, pos, audioPriority);
-        }
-        
-        public (float, AudioSource) PlayClipAtPos(AudioResource resource,
-            float3 pos,
-            byte audioPriority = (byte)EAudioPriority.Average) {
-            var currentSource = AudioSources[_currentIndex];
 
-            currentSource.transform.localPosition = pos;
-            currentSource.resource = resource;
-            currentSource.priority = audioPriority;
-            currentSource.Play();
-
-            var lengthSeconds = currentSource.clip?.length ?? 0f;
-
-            _previousIndex = _currentIndex;
-            _currentIndex++;
-            _currentIndex %= (byte)AudioSources.Length;
-
-            return (lengthSeconds, currentSource);
-        }
-        
-        public (float, AudioSource) PlayClip(string id, byte audioPriority = (byte)EAudioPriority.Average) {
-            return PlayClipAtPos(id, float3.zero, audioPriority);
-        }
-
-        public (float, AudioSource) PlayClip(AudioResource resource, byte audioPriority = (byte)EAudioPriority.Average) {
-            return PlayClipAtPos(resource, float3.zero, audioPriority);
-        }
-        
-        public bool StopClip(int idx = -1) {
-            idx = idx == -1 ? _previousIndex : idx;
-            var source = AudioSources[idx];
-
-            if (!source.isPlaying) return false;
-            source.Stop();
-
-            return true;
-        }
-
-        public bool PauseAllClips(bool isStop = false) {
-            PauseIndexes = ListPool<int>.Get();
-
-            for (var i = 0; i < AudioSources.Length; i++) {
-                var currentSource = AudioSources[i];
-
-                if (!currentSource.isPlaying) continue;
-
-                if (!isStop)
-                    currentSource.Pause();
-                else
-                    currentSource.Stop();
-                PauseIndexes.Add(i);
-            }
-
-            return PauseIndexes.Count == AudioSources.Length;
-        }
-
-        public bool ResumeClips() {
-            if (PauseIndexes is null) return false;
-
-            foreach (var index in PauseIndexes)
-                AudioSources[index].Play();
-
-            ListPool<int>.Release(PauseIndexes);
-
-            return true;
-        }
-
-        public abstract class Attribute : ScriptableObject {
-
-            public Reserve<AudioResource> pools;
-
-        }
     }
 
 }